// web/client/types.gen.ts
export type MessageOut = {
    /**
     * Content
     */
    content: string;
    role: Role;
    /**
     * File Name
     */
    file_name: string | null;
    /**
     * Chat Id
     */
    chat_id: string;
    /**
     * Id
     */
    id: number;
    /**
     * Created At
     */
    created_at: string;
};

/**
 * Role
 */
export type Role = 'user' | 'assistant' | 'system';